+++ src/components/Folder/Menus.js
@@ -1,4 +1,5 @@
 import { useState } from "react";
+// import { getFolderList } from "./../../api/api";
 import GlobalStyle from "./../common/GlobalStyle";
 import styled from "styled-components";
 import union from "../../assets/icons/Union.svg";
